requirements.txt
django-oauth-toolkit>=3.0.1
djangorestframework>=3.16.1
django-environ>=0.11.2